% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format_duration}
\alias{format_duration}
\title{Format durations with human friendly units}
\usage{
format_duration(secs)
}
\arguments{
\item{secs}{A numeric vector of durations in seconds.}
}
\value{
A character vector the same length as \code{secs}.
}
\description{
Chooses microseconds, milliseconds, seconds, or minutes and seconds based on
the size of each duration, and rounds to three significant digits.
}
\examples{
format_duration(c(0.0000123, 0.0456, 1.234, 75, 4000))
}
